Update temporary fake check
