+++ packages/apollo-shared/src/Checks/index.ts
@@ -1,0 +1,1 @@
+export * from './Checks'
