+++ packages/apollo-shared/src/index.ts
@@ -2,3 +2,4 @@ export * from './Changes'
 export * from './Operations'
 export * from './Validations'
 export * from './Common'
+export * from './Checks'
