+++ packages/jbrowse-plugin-apollo/src/BackendDrivers/InMemoryFileDriver.ts
@@ -1,30 +1,13 @@
 import { getConf } from '@jbrowse/core/configuration'
 import { Region, getSession } from '@jbrowse/core/util'
 import { AssemblySpecificChange, Change } from 'apollo-common'
-import { AnnotationFeatureSnapshot, CheckResultSnapshot } from 'apollo-mst'
+import { AnnotationFeatureSnapshot } from 'apollo-mst'
 import { ValidationResultSet } from 'apollo-shared'
-import { Check } from 'apollo-shared/src/Checks/Checks'
 
 import { SubmitOpts } from '../ChangeManager'
 import { BackendDriver } from './BackendDriver'
 
-export class InMemoryFileDriver extends BackendDriver implements Check {
-  async checkFeature(
-    feature: AnnotationFeatureSnapshot,
-  ): Promise<CheckResultSnapshot> {
-    const { _id, end, refSeq, start } = feature
-    const id = _id.toString()
-    return {
-      _id: `${id}-fake`,
-      name: 'FakeInMemoryCheckResult',
-      ids: [id],
-      refSeq: refSeq.toString(),
-      start,
-      end,
-      message: `This is a fake result for feature ${id}`,
-    }
-  }
-
+export class InMemoryFileDriver extends BackendDriver {
   async getFeatures(): Promise<AnnotationFeatureSnapshot[]> {
     return []
   }
